+++ src/components/Timeline/TimelineStage.js
