+++ src/components/Timeline/__tests__/MenuPanel.js
