--- src/containers/MainMenu/index.js
@@ -1,2 +1,0 @@
-export { default as SettingsMenu } from './SettingsMenu/SettingsMenu';
-export { default as StagesMenu } from './StagesMenu/StagesMenu';
