+++ src/containers/SettingsMenu/Sections/DeviceSettings.js
