+++ src/containers/SettingsMenu/Sections/VisualPreferences.js
