+++ src/containers/SettingsMenu/index.js
@@ -1,0 +1,1 @@
+export { default as SettingsMenu } from './SettingsMenu';
