+++ src/containers/Timeline/index.js
@@ -1,0 +1,1 @@
+export { default as StagesMenu } from './StagesMenu';
