--- src/styles/components/_main-menu.scss
@@ -1,6 +1,0 @@
-@import 'main-menu/main-menu';
-@import 'main-menu/settings-menu';
-@import 'main-menu/settings-menu-button';
-@import 'main-menu/stages-menu';
-@import 'main-menu/timeline';
-@import 'main-menu/timeline-stage';
