+++ src/styles/components/_settings-menu.scss
@@ -1,0 +1,3 @@
+@import 'settings-menu/settings-menu';
+@import 'settings-menu/settings-menu-button';
+@import 'settings-menu/select';
