+++ src/styles/components/_timeline.scss
@@ -1,110 +1,3 @@
-
-$timeline-button-size: 6rem;
-
-.timeline {
-  --light-background: #4a4677;
-  background: var(--background);
-  align-items: flex-end;
-  display: flex;
-  flex-direction: column;
-  position: absolute;
-  top: 0;
-  bottom: 0;
-  // left: -15rem;
-  z-index: var(--z-global-ui);
-  // border-radius: 0 7rem 7rem 0rem;
-  background: var(--panel-bg-muted);
-  padding: unit(2);
-  overflow: hidden;
-
-  .timeline-content {
-    display: flex;
-    padding: unit(2) 0;
-    flex-direction: column;
-    flex: 0 0 var(--timeline-width);
-    justify-content: center;
-    align-content: center;
-    align-items: center;
-    height: 100%;
-  }
-
-  svg {
-    height: 2rem;
-    width: 2rem;
-  }
-
-  &-nav {
-    align-items: center;
-    border-radius: 100%;
-    cursor: pointer;
-    display: flex;
-    justify-content: center;
-    height: $timeline-button-size;
-    width: $timeline-button-size;
-
-    &--next {
-      background: var(--light-background);
-
-      svg {
-        height: 2.5rem;
-        width: 2.5rem;
-      }
-
-      &::after {
-        content: '';
-        opacity: 0;
-        transition: transform var(--animation-duration-fast) var(--animation-easing), opacity var(--animation-duration-fast) var(--animation-easing);
-        display: block;
-        pointer-events: none;
-        position: absolute;
-        height: 9rem;
-        width: 9rem;
-        border-radius: 50%;
-        z-index: -1;
-        background: var(--primary);
-      }
-
-      &:active {
-        background: var(--primary);
-
-        &::after {
-          transform: scale(.5);
-          opacity: 1;
-          transition: 0s;
-        }
-      }
-    }
-
-    &--back {
-
-      .cls-2 {
-        fill: var(--transparent-light);
-      }
-
-      &::after {
-        content: '';
-        opacity: 0;
-        transition: transform var(--animation-duration-fast) var(--animation-easing), opacity var(--animation-duration-fast) var(--animation-easing);
-        display: block;
-        pointer-events: none;
-        position: absolute;
-        height: $timeline-button-size * 2;
-        width: $timeline-button-size * 2;
-        border-radius: 50%;
-        z-index: -1;
-        background: var(--light-background);
-      }
-
-      &:active {
-        background: var(--light-background);
-
-        &::after {
-          transform: scale(.5);
-          opacity: 1;
-          transition: 0s;
-
-        }
-      }
-    }
-  }
-}
+@import 'timeline/timeline';
+@import 'timeline/timeline-stage';
+@import 'timeline/stages-menu';
