--- src/styles/components/main-menu/_timeline.scss
@@ -1,15 +1,0 @@
-.menu-timeline {
-  flex: 1 1 auto;
-  overflow: hidden;
-  display: flex;
-  width: 26rem;
-  justify-content: center;
-
-  .scrollable {
-    padding-top: unit(2);
-  }
-
-  h4 {
-    align-self: center;
-  }
-}
