+++ src/styles/components/settings-menu/_select.scss
@@ -1,0 +1,33 @@
+.select-css {
+  display: block;
+  font-size: var(--font-size);
+  font-family: var(--body-font-family);
+  padding: unit(1);
+  width: 100%;
+  max-width: 100%;
+  height: 3.5rem;
+  box-sizing: border-box;
+  margin: 0;
+  border: 0;
+  border-radius: var(--border-radius);
+  -webkit-appearance: none;
+  appearance: none;
+  background-color: var(--color-platinum);
+  background-image: url('data:image/svg+xml;charset=US-ASCII,%3Csvg%20xmlns%3D%22http%3A%2F%2Fwww.w3.org%2F2000%2Fsvg%22%20width%3D%22292.4%22%20height%3D%22292.4%22%3E%3Cpath%20fill%3D%22%23007CB2%22%20d%3D%22M287%2069.4a17.6%2017.6%200%200%200-13-5.4H18.4c-5%200-9.3%201.8-12.9%205.4A17.6%2017.6%200%200%200%200%2082.2c0%205%201.8%209.3%205.4%2012.9l128%20127.9c3.6%203.6%207.8%205.4%2012.8%205.4s9.2-1.8%2012.8-5.4L287%2095c3.5-3.5%205.4-7.8%205.4-12.8%200-5-1.9-9.2-5.5-12.8z%22%2F%3E%3C%2Fsvg%3E');
+  background-repeat: no-repeat, repeat;
+  background-position: right .7em top 50%, 0 0;
+  background-size: .65em auto, 100%;
+  min-width: 6rem;
+}
+
+.select-css:hover {
+  border-color: var(--color-platinum--dark);
+}
+.select-css:focus {
+  border-color: var(--color-platinum--dark);
+}
+.select-css option {
+  font-weight:normal;
+}
+
+
