+++ src/styles/components/timeline/_stages-menu.scss
