+++ src/styles/components/timeline/_timeline-stage.scss
